Move the database view
